Extract identical code into shared function
